requirements.txt
mistralai
streamlit 
python-dotenv
ruff